bridges/mcp/pyproject.toml
[project]
name = "anip-mcp-bridge"
version = "0.1.0"
description = "Generic MCP bridge for any ANIP-compliant service"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
]

[project.scripts]
anip-mcp-bridge = "anip_mcp_bridge.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
